Account for counter rollover
In the case that a cgroups CPU time counter rolls over, use the previous
CPU percentage calculation. There should not be negative percentages.

This also fixes a bug where in the situation layed out above, the
calculations with unsigned integers could result in the high order
bits flipping to 1, causing inaccurate spikes in the result.